requirements.txt
numpy 
pandas 
scikit-learn 
torch 
transformers 
sentence-transformers 
faiss-cpu 
matplotlib 
scipy